Move async credentials stuff out to S3AsyncClientWorker
